+++ example/logic_test.sh
@@ -30,7 +30,7 @@ function test_should_validate_an_ok_exit_code() {
 
   fake_function
 
-  assertExitCode "0"
+  assert_exit_code "0"
 }
 
 
@@ -41,7 +41,7 @@ function test_should_validate_a_non_ok_exit_code() {
 
   fake_function
 
-  assertExitCode "1"
+  assert_exit_code "1"
 }
 
 function test_other_way_of_using_the_exit_code() {
@@ -49,7 +49,7 @@ function test_other_way_of_using_the_exit_code() {
     return 1
   }
 
-  assertExitCode "1" "$(fake_function)"
+  assert_exit_code "1" "$(fake_function)"
 }
 
 function test_successful_exit_code() {