+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   - `assertNotContains` -> `assert_not_contains`
   - `assertMatches` -> `assert_matches`
   - `assertNotMatches` -> `assert_not_matches`
+  - `assertExitCode` -> `assert_exit_code`
 
 ## [0.6.0](https://github.com/TypedDevs/bashunit/compare/0.5.0...0.6.0) - 2023-09-19